+++ tommath_class.h
@@ -42,7 +42,6 @@
 #   define BN_MP_EXCH_C
 #   define BN_MP_EXPORT_C
 #   define BN_MP_EXPT_D_C
-#   define BN_MP_EXPT_D_EX_C
 #   define BN_MP_EXPTMOD_C
 #   define BN_MP_EXTEUCLID_C
 #   define BN_MP_FREAD_C
@@ -77,7 +76,6 @@
 #   define BN_MP_MUL_D_C
 #   define BN_MP_MULMOD_C
 #   define BN_MP_N_ROOT_C
-#   define BN_MP_N_ROOT_EX_C
 #   define BN_MP_NEG_C
 #   define BN_MP_OR_C
 #   define BN_MP_PRIME_FERMAT_C
@@ -225,6 +223,10 @@
 #   define BN_MP_GET_LONG_LONG_C
 #   define BN_MP_PRIME_IS_DIVISIBLE_C
 #   define BN_S_MP_PRIME_IS_DIVISIBLE_C
+#   define BN_MP_EXPT_D_EX_C
+#   define BN_MP_EXPT_D_C
+#   define BN_MP_N_ROOT_EX_C
+#   define BN_MP_N_ROOT_C
 #endif
 
 #if defined(BN_MP_2EXPT_C)
@@ -384,10 +386,6 @@
 #endif
 
 #if defined(BN_MP_EXPT_D_C)
-#   define BN_MP_EXPT_D_EX_C
-#endif
-
-#if defined(BN_MP_EXPT_D_EX_C)
 #   define BN_MP_INIT_COPY_C
 #   define BN_MP_SET_C
 #   define BN_MP_MUL_C
@@ -616,16 +614,12 @@
 #endif
 
 #if defined(BN_MP_N_ROOT_C)
-#   define BN_MP_N_ROOT_EX_C
-#endif
-
-#if defined(BN_MP_N_ROOT_EX_C)
 #   define BN_MP_INIT_MULTI_C
 #   define BN_MP_COUNT_BITS_C
 #   define BN_MP_SET_C
 #   define BN_MP_2EXPT_C
 #   define BN_MP_COPY_C
-#   define BN_MP_EXPT_D_EX_C
+#   define BN_MP_EXPT_D_C
 #   define BN_MP_MUL_C
 #   define BN_MP_SUB_C
 #   define BN_MP_MUL_D_C
